+++ src/game/effect_support.rs
@@ -909,6 +909,8 @@ impl Game {
             | ObjectPredicateDef::PowerExactly(_)
             | ObjectPredicateDef::ToughnessExactly(_)
             | ObjectPredicateDef::ToughnessLessThan(_)
+            | ObjectPredicateDef::PowerGreaterThan(_)
+            | ObjectPredicateDef::ToughnessGreaterThan(_)
             | ObjectPredicateDef::ControlledBy(_)
             | ObjectPredicateDef::Supertype(_)
             | ObjectPredicateDef::DebutSet(_)