+++ src/game/tests.rs
@@ -391,6 +391,7 @@ mod decisions_and_effects;
 mod delayed_triggers;
 mod detain;
 mod entry_replacements;
+mod evolve;
 mod exalted;
 mod exile_source_costs;
 mod extra_turns;